+++ CHANGELOG.md
@@ -9,6 +9,13 @@ Fixed|Added|Changed|Deprecated|Removed|Security
 
     - Description of the change.
 
+## [0.1.0] - 2023-08-23
+
+### Added
+    - Added `remove_file_handler` method to `LogManager` to remove the file handler from the logger.
+    - Improved handling of log file preparation to avoid duplicate log entries.
+    - Added comments and documentation for better code clarity.
+
 ## [0.0.2] - 2023-07-10
 
 ### Added